Reduce size of cache variables
The variables could explode for huge directory trees. Now when searching
through a relevant directory, just the relevant names will fill the
variables. This should further improve speed.
There is still places to improve with this approach. Some "find /" were
left untouched because they seem to be non-trivial searches filtering
permission bits, for example.
There is a huge bottleneck in the ssh section, namely the search for the
privatekeyfiles! That's why I shifted this process nearer to its use.